journald: log volatile HYP-9051
So far, the log messages are continuously written into persistent NAND memory.
Now the log messages are written into the RAM, to avoid an early NAND memory
aging.
The second reason to write log messages into RAM is that sometimes the rotation
of the journald log files does not work. In this case the NAND memory was filled
up to 100% with log messages.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>