hipos-device-info.sh: bug fixing and code review
The part of the code creating '/etc/hipos/hipos-device.<type>' (whose
existence is then used as a condition in systemd services) is now always
executed, even if we cannot access the board controller.

If no device type is set in the device file, 'unknown' is used. That is,
the file '/etc/hipos/hipos-device.unknown' will be created

Handling of several other errors has also been improved and made more
consistent.

Signed-off-by: Angelika Tobisch <tobisch@dresearch-fe.de>
Acked-by: Ralf Schröder <schroeder@dresearch-fe.de>