ldb: lvds power up/down sequence
This commit implements a lvds power up and down sequence.
Many lvds displays have a power sequence as follows:

             ___________   _
VCC        _/           \_/
                _____
Data       ____|     |_____
                  _
Backlight _______| |_______

There is a minimum and maximum time requirement between VCC and data and
between data and backlight. To fullfill the time requirements, the ldb driver
needs to know VCC and backlight GPIO and the timing before and after data
enable. Now this can be configured using following "lvds-channel" device tree
elements:

&ldb {
        ...
        lvds-channel@0 {
                ...
                vcc-gpio = <&gpio4 7 GPIO_ACTIVE_LOW>;
                vcc-powerup-delay-ms = <1>;
                backlight-gpio = <&gpio4 6 GPIO_ACTIVE_HIGH>;
                backlight-powerup-delay-ms = <200>;
                ...
        };
};

"vcc-gpio" is the VCC GPIO.
"vcc-powerup-delay-ms" is the delay between VCC and data in milliseconds during
powerup.
"backlight-gpio" is the backlight GPIO.
"backlight-powerup-delay-ms" is the delay between data and backlight in
milliseconds during powerup.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>