mmc: sdhci: recompute timeout_clk HYP-13485
This is a backport of mmc sdhci timeout_clk calculation.

When SDHCI_QUIRK_DATA_TIMEOUT_USES_SDCLK is set, timeout_clk is sdclk.
We need to update it when we change sdclk in sdhci_set_clock.
This allow to have a more precisse timeout and max_busy_timeout. This
can help for command that need a big busy wait (erase, ...)

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>