imx-sdma: Update channel HYP-14823
Sometimes no data are read from a serial port. For example, this behavior occurs
when reading out GPS data. At least in one of 10 cases, no data are read.
This commit updates the DMA channels inside the IRQ handler. This is the same
behavior as in Kernel 3.14.28. And the error described no longer occurs.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>