IVAP u-boot for DVREC usable HYP-16664
The same u-boot can be used for IVAP and DVREC boards. The pad ENET_RX_ER can
be used to determine whether it is an IVAP or DVREC board. The DVREC board has
a pulldown and the IVAP board does not. Is the value 0 then it is a DVREC board
otherwise it is an IVAP board.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>