Add real tar to image HYP-19186
With yocto sumo branch the busybox tar reports following error message
when unpacking the rootfs:

tar: skipping unsafe symlink to '../run/lock' in archive, set EXTRACT_UNSAFE_SYMLINKS=1 to extract

A following boot from this rootfs is not possible. The real tar does not
have this problems. Therefore tar and bzip2 are installed.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>