hinat: Update kernel configuration HYP-23856
Update defconfig to kernel 5.4.
Copy defconfig in configure step. As the linux-yocto_5.4.bb recipe is
extended, the kernel metadata mechanism, which configures the kernel
from a git repository, is also applied. These leads to an overwriting of
kernel configuration values from defconfig. To avoid this overwriting,
the defconfig, which is a full kernel configuration, is copied after the
do_kernel_configme step.

Signed-off-by: Mario Schuknecht <mario.schuknecht@dresearch-fe.de>