libsoup-2.4: Consider max size when reading websocket data, HYP-32526
One can set the maximum payload size with soup_websocket_connection_set_max_incoming_payload_size().
But apparently this is only checked against the header data of a packet and not the data
actually sent.
Added patch to check in soup_websocket_connection_read() that the size is not exceeded.
FYI: Default is 128kb

Signed-off-by: Andreas Zisowsky <zisowsky@dresearch-fe.de>